refactor: remove unused authentication actions and controllers
