--- app/Actions/Auth/LoginAction.php
@@ -1,82 +1,0 @@
-<?php
-
-namespace App\Actions\Auth;
-
-use App\Models\User;
-use Illuminate\Auth\Events\Lockout;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\RateLimiter;
-use Illuminate\Support\Facades\Validator;
-use Illuminate\Support\Str;
-use Illuminate\Validation\ValidationException;
-
-class LoginAction
-{
-    public function execute(string $email, string $password, bool $remember = false, ?string $ip = null): User
-    {
-        $validator = Validator::make([
-            'email' => $email,
-            'password' => $password,
-        ], [
-            'email' => 'required|email',
-            'password' => 'required|string',
-        ]);
-
-        if ($validator->fails()) {
-            throw new ValidationException($validator);
-        }
-
-        $this->ensureIsNotRateLimited($email, $ip);
-
-        $credentials = [
-            'email' => strtolower($email),
-            'password' => $password,
-        ];
-
-        if (! Auth::attempt($credentials, $remember)) {
-            RateLimiter::hit($this->throttleKey($email, $ip));
-
-            throw ValidationException::withMessages([
-                'email' => [__('auth.failed')],
-            ]);
-        }
-
-        RateLimiter::clear($this->throttleKey($email, $ip));
-
-        return Auth::user();
-    }
-
-    /**
-     * Ensure the login request is not rate limited.
-     *
-     * @throws ValidationException
-     */
-    private function ensureIsNotRateLimited(string $email, ?string $ip): void
-    {
-        if (! RateLimiter::tooManyAttempts(
-            $this->throttleKey($email, $ip),
-            config('auth.login_rate_limit', 5)
-        )) {
-            return;
-        }
-
-        event(new Lockout(request()));
-
-        $seconds = RateLimiter::availableIn($this->throttleKey($email, $ip));
-
-        throw ValidationException::withMessages([
-            'email' => [__('auth.throttle', [
-                'seconds' => $seconds,
-                'minutes' => ceil($seconds / 60),
-            ])],
-        ]);
-    }
-
-    /**
-     * Get the rate limiting throttle key for the request.
-     */
-    private function throttleKey(string $email, ?string $ip): string
-    {
-        return Str::transliterate(Str::lower($email).'|'.($ip ?? request()->ip()));
-    }
-}
