--- app/Actions/Auth/LogoutAction.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace App\Actions\Auth;
-
-use Illuminate\Support\Facades\Auth;
-
-class LogoutAction
-{
-    public function execute(): void
-    {
-        Auth::guard('web')->logout();
-    }
-}
