--- app/Actions/Auth/RegisterUserAction.php
@@ -1,43 +1,0 @@
-<?php
-
-namespace App\Actions\Auth;
-
-use App\Models\User;
-use Illuminate\Auth\Events\Registered;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Facades\Validator;
-use Illuminate\Validation\Rules\Password;
-use Illuminate\Validation\ValidationException;
-
-class RegisterUserAction
-{
-    public function execute(string $name, string $email, string $password): User
-    {
-        $validator = Validator::make([
-            'name' => $name,
-            'email' => $email,
-            'password' => $password,
-        ], [
-            'name' => 'required|string|max:255',
-            'email' => 'required|string|email|max:255|unique:'.User::class,
-            'password' => ['required', Password::defaults()],
-        ]);
-
-        if ($validator->fails()) {
-            throw new ValidationException($validator);
-        }
-
-        $user = User::create([
-            'name' => $name,
-            'email' => $email,
-            'password' => Hash::make($password),
-        ]);
-
-        Auth::login($user);
-
-        event(new Registered($user));
-
-        return $user;
-    }
-}
