--- app/Actions/Social/GetEnabledProvidersAction.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace App\Actions\Social;
-
-class GetEnabledProvidersAction
-{
-    public function execute(): array
-    {
-        return collect(config('app.social_providers', []))
-            ->filter(fn ($config) => ($config['enabled'] ?? false) === true)
-            ->map(fn ($config) => ['name' => $config['name']])
-            ->toArray();
-    }
-}
