--- app/Http/Controllers/Auth/AuthenticatedSessionController.php
@@ -1,62 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Auth;
-
-use App\Actions\Auth\LoginAction;
-use App\Actions\Auth\LogoutAction;
-use App\Http\Controllers\Controller;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Route;
-use Illuminate\Validation\ValidationException;
-use Inertia\Inertia;
-use Inertia\Response;
-
-class AuthenticatedSessionController extends Controller
-{
-    /**
-     * Display the login view.
-     */
-    public function create(): Response
-    {
-        return Inertia::render('auth/Login', [
-            'canResetPassword' => Route::has('password.request'),
-            'status' => session('status'),
-        ]);
-    }
-
-    /**
-     * Handle an incoming authentication request.
-     */
-    public function store(Request $request, LoginAction $loginAction): RedirectResponse
-    {
-        try {
-            $loginAction->execute(
-                $request->email,
-                $request->password,
-                $request->boolean('remember'),
-                $request->ip()
-            );
-
-            $request->session()->regenerate();
-
-            return redirect()->intended(route('dashboard', absolute: false));
-        } catch (ValidationException $e) {
-            throw $e;
-        }
-    }
-
-    /**
-     * Destroy an authenticated session.
-     */
-    public function destroy(Request $request, LogoutAction $logoutAction): RedirectResponse
-    {
-        $logoutAction->execute();
-
-        $request->session()->invalidate();
-
-        $request->session()->regenerateToken();
-
-        return redirect('/');
-    }
-}
