--- app/Http/Controllers/Auth/PasswordController.php
@@ -1,57 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Auth;
-
-use App\Actions\Auth\ResetPasswordAction;
-use App\Http\Controllers\Controller;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Validation\Rules\Password;
-use Illuminate\Validation\ValidationException;
-use Inertia\Inertia;
-use Inertia\Response;
-
-class PasswordController extends Controller
-{
-    /**
-     * Display the password reset view.
-     */
-    public function create(Request $request): Response
-    {
-        return Inertia::render('auth/ResetPassword', [
-            'email' => $request->email,
-            'token' => $request->route('token'),
-        ]);
-    }
-
-    /**
-     * Handle an incoming password reset request.
-     *
-     * @throws ValidationException
-     */
-    public function store(Request $request, ResetPasswordAction $resetAction): RedirectResponse
-    {
-        $resetAction->execute($request->email, $request->password, $request->token);
-
-        return redirect()->route('login')->with('status', __('passwords.reset'));
-    }
-
-    /**
-     * Update the authenticated user's password.
-     */
-    public function update(Request $request): RedirectResponse
-    {
-        $validated = $request->validate([
-            'current_password' => ['required', 'current_password'],
-            'password' => ['required', Password::defaults(), 'confirmed'],
-        ]);
-
-        Auth::user()->update([
-            'password' => Hash::make($validated['password']),
-        ]);
-
-        return redirect()->route('profile.edit');
-    }
-}
