--- app/Http/Controllers/Auth/RegisteredUserController.php
@@ -1,34 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\Auth;
-
-use App\Actions\Auth\RegisterUserAction;
-use App\Http\Controllers\Controller;
-use Illuminate\Http\RedirectResponse;
-use Illuminate\Http\Request;
-use Illuminate\Validation\ValidationException;
-use Inertia\Inertia;
-use Inertia\Response;
-
-class RegisteredUserController extends Controller
-{
-    /**
-     * Display the registration view.
-     */
-    public function create(): Response
-    {
-        return Inertia::render('auth/Register');
-    }
-
-    /**
-     * Handle an incoming registration request.
-     *
-     * @throws ValidationException
-     */
-    public function store(Request $request, RegisterUserAction $registerAction): RedirectResponse
-    {
-        $registerAction->execute($request->name, $request->email, $request->password);
-
-        return redirect(route('dashboard', absolute: false));
-    }
-}
